Currently in a buggy state
We appear to be introducing duplicate edges in the swap_edges function but I am not sure how yet.

Label: Coreg. Adding Reg edge: WBGene00003163 -> WBGene00001568

Label: Reg. Adding Reg edge: WBGene00003163 -> WBGene00001568

This is an example of when the same edge has multiple labels. I don't quite know why this is happening but I think it is either in the swap_edges, label_edges or split_to_csv functions